add render_common_button

+++ wxFEFactory/python/tools/gta_base/utils.py
@@ -2,7 +2,7 @@ import math
 PI = math.pi
 
 def degreeToRadian(degrees):
-    return degrees * float(PI / 180.0)
+    return degrees * (PI / 180.0)
 
 def headingToDirection(heading):
     heading = degreeToRadian(heading)
